+++ web/templates/index.html
@@ -1,0 +1,64 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Disease Prediction Model</title>
+    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.1/css/all.min.css">
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet"
+        integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
+</head>
+
+<body class="main-screen">
+    <div class="main-card">
+        <div class="logo-row">
+            <div class="nav-logo">
+                <a href="LOGO PNG ALS.png">
+                    <img src="https://i.ibb.co/FDK32F4/LOGO-PNG-ALS.png" alt="Web Logo" style="height: 50px; width: auto;">
+                </a>
+            </div>
+        </div>
+        <div class="main-content">
+            <!-- Left Section -->
+            <div class="intro-section">
+                <h1 class="model-name">Disease Prediction Model</h1>
+                <p class="model-intro">
+                    Lorem ipsum dolor sit amet consectetur, adipisicing elit. Cum eum expedita numquam ducimus fugiat ab?
+                    Harum sequi architecto quaerat quis quo adipisicing elit. Cum eum expedita numquam ducimus fugiat ab?
+                    Harum sequi architecto quaerat quis quo odioisicing elit. Cum eum expedita numquam ducimus fugiat ab?
+                    Harum sequi architecto quaerat quis quo odioisicing elit. Cum eum expedita numquam ducimus fugiat ab?
+                    Harum sequi architecto quaerat quis quo odio, reiciendis blanditiis provident fugit esse nam,
+                    quidem voluptate?
+                </p>
+                <div class="logo">
+                    <a href="https://www.facebook.com" target="_blank"><i class="fa-brands fa-facebook-f"></i></a>
+                    <a href="https://www.instagram.com" target="_blank"><i class="fa-brands fa-instagram"></i></a>
+                    <a href="https://www.github.com" target="_blank"><i class="fa-brands fa-github"></i></a>
+                </div>
+            </div>
+
+            <!-- Right Section -->
+            <div class="model-input">
+                <input type="text" name="User-Name" id="name-coloumn" placeholder="Your Name..." required>
+                <div class="symptoms-section">
+                    <p>What are your symptoms?</p>
+                    <div class="symptom-tags" id="symp-tags">
+                    </div>  
+                    <select class="form-select mt-2" id="symp-select">
+                        <option selected>Add a symptom...</option>
+                    </select>                    
+                </div>
+                <button type="button" class="btn btn-primary mt-3" onclick="userInfo()" >Predict Disease</button>
+            </div>
+        </div>
+    </div>
+
+    <script src="app.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
+        integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz" crossorigin="anonymous"></script>
+</body>
+
+</html>
